refactor: owner 패키지 분리
