+++ src/main/java/in/koreatech/koin/admin/user/controller/AdminUserController.java
@@ -4,11 +4,9 @@ import static in.koreatech.koin.domain.user.model.UserType.ADMIN;
 
 import java.net.URI;
 
-import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -18,11 +16,6 @@ import org.springframework.web.bind.annotation.RestController;
 
 import in.koreatech.koin.admin.user.dto.AdminLoginRequest;
 import in.koreatech.koin.admin.user.dto.AdminLoginResponse;
-import in.koreatech.koin.admin.user.dto.AdminNewOwnersResponse;
-import in.koreatech.koin.admin.user.dto.AdminOwnerResponse;
-import in.koreatech.koin.admin.user.dto.AdminOwnerUpdateRequest;
-import in.koreatech.koin.admin.user.dto.AdminOwnerUpdateResponse;
-import in.koreatech.koin.admin.user.dto.AdminOwnersResponse;
 import in.koreatech.koin.admin.user.dto.AdminPasswordChangeRequest;
 import in.koreatech.koin.admin.user.dto.AdminPermissionUpdateRequest;
 import in.koreatech.koin.admin.user.dto.AdminResponse;
@@ -32,7 +25,6 @@ import in.koreatech.koin.admin.user.dto.AdminUpdateRequest;
 import in.koreatech.koin.admin.user.dto.AdminsCondition;
 import in.koreatech.koin.admin.user.dto.AdminsResponse;
 import in.koreatech.koin.admin.user.dto.CreateAdminRequest;
-import in.koreatech.koin.admin.user.dto.OwnersCondition;
 import in.koreatech.koin.admin.user.enums.TeamType;
 import in.koreatech.koin.admin.user.enums.TrackType;
 import in.koreatech.koin.admin.user.service.AdminUserService;
@@ -46,14 +38,6 @@ import lombok.RequiredArgsConstructor;
 public class AdminUserController implements AdminUserApi{
 
     private final AdminUserService adminUserService;
-
-    @PutMapping("/admin/owner/{id}/authed")
-    public ResponseEntity<Void> allowOwnerPermission(
-        @PathVariable Integer id,
-        @Auth(permit = {ADMIN}) Integer adminId) {
-        adminUserService.allowOwnerPermission(id);
-        return ResponseEntity.ok().build();
-    }
 
     @PostMapping("/admin")
     public ResponseEntity<Void> createAdmin(
@@ -157,41 +141,6 @@ public class AdminUserController implements AdminUserApi{
     ) {
         adminUserService.updateAdminPermission(request, id, adminId);
         return ResponseEntity.ok().build();
-    }
-
-    @GetMapping("/admin/users/owner/{id}")
-    public ResponseEntity<AdminOwnerResponse> getOwner(
-        @PathVariable Integer id,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        AdminOwnerResponse adminOwnerResponse = adminUserService.getOwner(id);
-        return ResponseEntity.ok().body(adminOwnerResponse);
-    }
-
-    @PutMapping("/admin/users/owner/{id}")
-    public ResponseEntity<AdminOwnerUpdateResponse> updateOwner(
-        @PathVariable Integer id,
-        @RequestBody @Valid AdminOwnerUpdateRequest request,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        AdminOwnerUpdateResponse adminOwnerUpdateResponse = adminUserService.updateOwner(id, request);
-        return ResponseEntity.ok().body(adminOwnerUpdateResponse);
-    }
-
-    @GetMapping("/admin/users/new-owners")
-    public ResponseEntity<AdminNewOwnersResponse> getNewOwners(
-        @ParameterObject @ModelAttribute OwnersCondition ownersCondition,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        return ResponseEntity.ok().body(adminUserService.getNewOwners(ownersCondition));
-    }
-
-    @GetMapping("/admin/users/owners")
-    public ResponseEntity<AdminOwnersResponse> getOwners(
-        @ParameterObject @ModelAttribute OwnersCondition ownersCondition,
-        @Auth(permit = {ADMIN}) Integer adminId
-    ) {
-        return ResponseEntity.ok().body(adminUserService.getOwners(ownersCondition));
     }
 
     @GetMapping("/admin/users/{id}")
